requirements.txt
asgiref==3.7.2
Django==3.2.13
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
